+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 **Internal**:
 
-- Add capability to configure metrics aggregators per use case. ([2341](https://github.com/getsentry/relay/pull/2341))
+- Add capability to configure metrics aggregators per use case. ([#2341](https://github.com/getsentry/relay/pull/2341))
 
 ## 23.7.0