+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Interpret `aggregator.max_tag_value_length` as characters instead of bytes. ([#2343](https://github.com/getsentry/relay/pull/2343))
+- Add capability to configure metrics aggregators per use case. ([#2341](https://github.com/getsentry/relay/pull/2341))
 
 ## 23.7.0